+++ src/plat/sys.unix.cpp
@@ -7,16 +7,13 @@
 #ifndef _WIN32
 
 #include <chen/sys/sys.hpp>
-#include <cstring>
-#include <cstdlib>
 #include <cerrno>
 
 // -----------------------------------------------------------------------------
 // sys
-std::string chen::sys::error()
+std::error_code chen::sys::error()
 {
-    char buf[1024] = {};
-    return !errno ? "No error" : sys::peek(::strerror_r(errno, buf, sizeof(buf)), buf);
+    return std::error_code(errno, std::system_category());
 }
 
 #endif